Unify pool tests in tests/, move rarely used helper function to the place of usage
